twai: do not abort transmissions on recoverable errors
The TWAI driver's interrupt handler currently aborts transmissions if
certain (error) interrupts were triggered. This causes the transmission
of frames to become unreliable, because the `transmit_async` operation
does not return any error codes if a transmission was aborted.
Additionally, the TWAI controller is designed to be a reliable and, with
the exception of the bus-off state, normally would simply try to
rentransmit a message on its own. Therefore, this commit removes
transmission abortion on errors and, instead, only aborts a transmission
if the controller changes into the bus-off state.